requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
torch==1.9.0
transformers==4.11.3
numpy==1.21.2
scipy==1.7.1
pydub==0.25.1
openvoice-xtts==2.0.0
python-multipart==0.0.5
ffmpeg-python==0.2.0 